# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Career Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ── Resumes ────────────────────────────────────────────────────────────────
...
  /roadmaps/generate:
    post:
      operationId: generateRoadmap
      tags: [roadmaps]
      summary: Generate a personalized learning roadmap based on skill gaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadmapGenerateRequest"
      responses:
        "201":
          description: Created roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningRoadmap"
  /roadmaps/{id}:
    get:
      operationId: getRoadmap
      tags: [roadmaps]
      summary: Get a learning roadmap
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningRoadmap"
        "404":
          description: Not found
    delete:
      operationId: deleteRoadmap
      tags: [roadmaps]
      summary: Delete a roadmap
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: number
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        targetRole:
          type: ["string", "null"]
        experienceLevel:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - createdAt
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        targetRole:
          type: string
        experienceLevel:
          type: string
        bio:
          type: string

    # ── Resumes ────────────────────────────────────────────────────────────
...
    LearningRoadmap:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        targetRole:
          type: string
        description:
          type: string
        totalSteps:
          type: number
        completedSteps:
          type: number
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapStep"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - targetRole
        - description
        - totalSteps
        - completedSteps
        - steps
        - createdAt
    RoadmapStep:
      type: object
      properties:
        id:
          type: number
        roadmapId:
          type: number
        title:
          type: string
        description:
          type: string
        resourceUrl:
          type: ["string", "null"]
        resourceType:
          type: string
        estimatedHours:
          type: ["number", "null"]
        orderIndex:
          type: number
        isCompleted:
          type: boolean
      required:
        - id
        - roadmapId
        - title
        - description
        - resourceType
        - orderIndex
        - isCompleted
    RoadmapGenerateRequest:
      type: object
      properties:
        targetRole:
          type: string
        currentSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        resumeId:
          type: ["number", "null"]
        jobId:
          type: ["number", "null"]
      required:
        - targetRole
...